requirements.txt
# Core Dependencies
pinecone-client>=3.0.0
sentence-transformers>=2.2.0
openai>=1.0.0
nltk>=3.8.1
pandas>=2.0.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.21.0
aiohttp>=3.8.1
async-timeout>=4.0.2
ratelimit>=3.0.0

# Progress bars
tqdm>=4.65.0

# GUI and web
tkinter>=8.6
flask>=2.3.2

# Data processing
numpy>=1.24.0

# Retrying
tenacity>=9.0.0
